ray tracing a red sphere successfully

+++ src/intersection_utils.c
@@ -6,7 +6,7 @@
 /*   By: mgaldino <mgaldino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/07 11:49:30 by mgaldino          #+#    #+#             */
-/*   Updated: 2022/12/12 09:47:49 by mgaldino         ###   ########.fr       */
+/*   Updated: 2022/12/12 12:56:24 by mgaldino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,11 @@ t_list	*get_intersections(float a, float b, float c, t_elements *sphere)
 	t_intersection	*new_intersection1;
 	t_intersection	*new_intersection2;
 	t_list			*intersections;
-	float			discriminant;
+	double			discriminant;
 
 	discriminant = pow(b, 2) - 4 * a * c;
+	//if (discriminant > 0)
+	//	printf("discriminant = %f\n", discriminant);
 	if (discriminant < 0)
 		return (NULL);
 	new_intersection1 = (t_intersection *) malloc(sizeof(t_intersection));
